chore: refactor find command
