+++ pkg/networking/networking.go
@@ -12,7 +12,7 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/sralloza/rpi-provisioner/pkg/info"
 	"github.com/sralloza/rpi-provisioner/pkg/logging"
-	"github.com/sralloza/rpi-provisioner/ssh"
+	"github.com/sralloza/rpi-provisioner/pkg/ssh"
 )
 
 type NetworkingArgs struct {
